Support all courant cmap subtables used on the Windows platform

+++ FontAnalyzer/TTF/Cmap/ICmapSubtable.cs
@@ -2,6 +2,6 @@
 {
     public interface ICmapSubtable
     {
-        ushort GetGlyphIndex(char c);
+        uint GetGlyphIndex(char c);
     }
 }
